Move hyp3 services into subfolder
