+++ src/app/services/hyp3/index.ts
@@ -1,0 +1,3 @@
+export * from './hyp3.service';
+export * from './on-demand.service';
+export * from './possible-hyp3-jobs.service';
