+++ src/app/services/hyp3/possible-hyp3-jobs.service.ts
